+++ web/firebase-messaging-sw.js
@@ -21,23 +21,8 @@ firebase.initializeApp({
 // Retrieve an instance of Firebase Messaging so that it can handle background
 // messages.
 const messaging = firebase.messaging();
-messaging.setBackgroundMessageHandler(function (payload) {
-    const promiseChain = clients
-        .matchAll({
-            type: "window",
-            includeUncontrolled: true
-        })
-        .then(windowClients => {
-            for (let i = 0; i < windowClients.length; i++) {
-                const windowClient = windowClients[i];
-                windowClient.postMessage(payload);
-            }
-        })
-        .then(() => {
-            return registration.showNotification("New Message");
-        });
-    return promiseChain;
-});
-self.addEventListener('notificationclick', function (event) {
-    console.log('notification received: ', event)
+
+// Optional:
+messaging.onBackgroundMessage((m) => {
+    console.log("onBackgroundMessage", m);
 });